Made integer types wider to attempt to calculate ack(3, 20) with memoization but failed :(

+++ c/ackermann.c
@@ -1,24 +1,26 @@
 #include "ackermann.h"
 
-void _cache_ack(hash_table *table, long value, pair key);
-long _memoized(int m, int n, hash_table *table);
+/* Allocates an 'ack' on the heap and associates it with a given key */
+void _cache_ack(hash_table *table, uint64_t value, pair key);
+/* Internal memoized ackermann calculation */
+uint64_t _memoized(uint64_t m, uint64_t n, hash_table *table);
 
-long naive(int m, int n) {
+uint64_t naive(uint64_t m, uint64_t n) {
   if (m == 0) 
-    return (long) n + 1;
+    return (uint64_t) n + 1;
   else if (n == 0) 
     return naive(m-1, 1);
   else
     return naive(m-1, naive(m, n-1));
 }
 
-long memoized(int m, int n) {
+uint64_t memoized(uint64_t m, uint64_t n) {
     hash_table *t = hash_table_create(100);
     if (t == NULL) {
       fprintf(stderr, "Failed to create hash table.");
       abort();
     }
-  long val = _memoized(m, n, t);
+  uint64_t val = _memoized(m, n, t);
   size_t i;
   for (i=0; i<t->size; i++) {
     free(t->table[i]);
@@ -27,28 +29,28 @@ long memoized(int m, int n) {
   return val;
 }
 
-long _memoized(int m, int n, hash_table *table) {
+uint64_t _memoized(uint64_t m, uint64_t n, hash_table *table) {
   pair key = {.m=m, .n=n};
   ack *cached = hash_table_lookup(table, key);
   if (cached != NULL) 
     return cached->value;
   if (m == 0) {
-    long value = (long) n + 1;
+    uint64_t value = (uint64_t) n + 1;
     _cache_ack(table, value, key);
     return value;
   } else if (n == 0) {
-    long value = _memoized(m-1, 1, table);
+    uint64_t value = _memoized(m-1, 1, table);
     _cache_ack(table, value, key);
     return value;
   }
   else {
-    long value = _memoized(m-1, _memoized(m, n-1, table), table);
+    uint64_t value = _memoized(m-1, _memoized(m, n-1, table), table);
     _cache_ack(table, value, key);
     return value;
   }
 }
 
-void _cache_ack(hash_table *table, long value, pair key) {
+void _cache_ack(hash_table *table, uint64_t value, pair key) {
     ack *cache = malloc(sizeof(ack));
     if (cache == NULL) {
       fprintf(stderr, "Memory allocation failed.\n");
@@ -64,19 +66,19 @@ void _cache_ack(hash_table *table, long value, pair key) {
     }
 }
 
-long iterative(int i, int n) {
-  long *next = calloc((long) i+1, sizeof(long));
-  long *goal = calloc((long) i+1, sizeof(long));
-  int idx = 0;
+uint64_t iterative(uint64_t i, uint64_t n) {
+  uint64_t *next = calloc((uint64_t) i+1, sizeof(uint64_t));
+  uint64_t *goal = calloc((uint64_t) i+1, sizeof(uint64_t));
+  uint64_t idx = 0;
   while (idx < i+1){
     goal[idx++] = 1;
   }
   goal[i] = -1;
-  long value = -1;
+  uint64_t value = -1;
   do  {
     value = next[0] + 1;
     bool transferring = true;
-    long current = i;
+    uint64_t current = i;
     while (transferring) {
       if (next[i-current] == goal[i-current]) {
         goal[i-current] = value;
@@ -87,7 +89,7 @@ long iterative(int i, int n) {
       current -= 1;
     }
   } 
-  while (next[i] != (long) n + 1);
+  while (next[i] != (uint64_t) n + 1);
   free(next);
   free(goal);
   return value;
