+++ c/libht/libht.c
@@ -9,7 +9,7 @@ int _pair_cmp(pair p1, pair p2);
   * Hashes 'key' for insertion in a hash_table with a capacity of 'size'
   * elements.
   */
-unsigned int _hash(pair key, size_t size); 
+uint64_t _hash(pair key, size_t size); 
 
 hash_table *hash_table_create(size_t size) {
   hash_table *ht;
@@ -25,13 +25,16 @@ hash_table *hash_table_create(size_t size) {
 }
 
 void hash_table_destroy(hash_table *ht) {
+  /* Since we make no assumptions about our clients data, freeing individual
+   * table entries is up to the user.
+   */
   free(ht->table);
   free(ht);
 }
 
 bool hash_table_insert(hash_table *ht, pair key, ack *value) {
   if (ht == NULL || value == NULL) return false;
-  int index = _hash(key, ht->size);
+  uint64_t index = _hash(key, ht->size);
   /* This currently rejects on collisions, but it should chain/probe instead.
   if (ht->table[index] != NULL) {
     return false;
@@ -42,7 +45,7 @@ bool hash_table_insert(hash_table *ht, pair key, ack *value) {
 
 ack *hash_table_delete(hash_table *ht, pair pair) {
   if (ht == NULL) return NULL;
-  int index = _hash(pair, ht->size);
+  uint64_t index = _hash(pair, ht->size);
   if (ht->table[index] != NULL &&
     _pair_cmp(ht->table[index]->pair, pair) == 0) {
     ack *tmp = ht->table[index];
@@ -53,7 +56,7 @@ ack *hash_table_delete(hash_table *ht, pair pair) {
 }
 
 ack *hash_table_lookup(hash_table * ht, pair pair) {
-  int index = _hash(pair, ht->size);
+  uint64_t index = _hash(pair, ht->size);
   if (ht->table[index] != NULL &&
     _pair_cmp(ht->table[index]->pair, pair) == 0) return ht->table[index];
   return NULL;
@@ -67,15 +70,15 @@ void print_hash_table(hash_table *ht) {
       printf("\t%ld\t---\n", i);
     } else {
       ack *a = ht->table[i];
-      printf("\t%ld\t---\t(%d, %d)=>%ld\t\n", i, a->pair.m, a->pair.n, a->value);
+      printf("\t%ld\t---\t(%ld, %ld)=>%ld\t\n", i, a->pair.m, a->pair.n, a->value);
     }
   }
   printf("\n");
 }
 
-unsigned int _hash(pair pair, size_t size) {
-  int m = pair.m;
-  int n = pair.n;
+uint64_t _hash(pair pair, size_t size) {
+  uint64_t m = pair.m;
+  uint64_t n = pair.n;
   return ((m*m + m + 2*m*n + 3*n + n*n) / 2) % size;
 }
